🔧 Make spec_helper to be required
This is needed to remove the need to  always write "require
spec_helper" for every new test file that is written.

The side effect is that as a developer, you won't have to always write
the line "require spec_helper" at the top of every test file that you
write.

Co-authored-by: Rob Whittaker <rob@thoughtbot.com>