🔧 Add honeybadger
Before, in case of an exception/error happening on the staging and
production server, we would have to use the terminal to identify the
root cause of the issue.

Honeybadger was added to make it easy and quick to track exceptions so
that the issue that has come up can be solved.