🚨 Add system test check if navigation is hidden
Before, we had a navigation menu that would always show on every page
including the sign in page.
This was rather destructive.

Now, we have a test that checks if the sign in link is on the sign in
page and if it is, we are hiding it together with the rest of the
navigation menu.