Update `.sample.env`
`.sample.env` is copied to `.env` when running `bin/setup`. However,
some environment variables are not set and then have to be added
manually.

This commit updates `.sample.env` with example values for the missing
environment variables to make it easier to set up this project after the
initial clone.